Connect Redux store to React app
