+++ src/redux/Books/actions.js
@@ -1,10 +1,11 @@
 import { bookAdded, bookRemoved } from './actionTypes';
 
-export const addBook = (bookTitle, author) => ({
+export const addBook = (bookTitle, author, topic) => ({
   type: bookAdded,
   payload: {
     bookTitle,
     author,
+    topic,
   },
 });
